requirements.txt
accelerate==0.28.0
beir==2.0.0
bm25s==0.1.10
codecarbon==2.6.0
cohere==5.9.1
evaluate==0.4.1
faiss_cpu==1.8.0.post1
flash_attn==2.5.7
google_cloud_aiplatform==1.49.0
GPUtil==1.4.0
gritlm==1.0.1
huggingface_hub==0.23.4
jax==0.4.31
lightning==2.4.0
nltk==3.8.1
numpy==2.1.1
openai==1.43.1
packaging==24.1
pandas==2.2.2
peft==0.12.0
psutil==5.9.8
psutil==5.9.0
pydantic==2.9.0
pyserini==0.36.0
PyStemmer==2.2.0.1
pytorch_metric_learning==2.6.0
pytrec_eval==0.5
pytrec_eval_terrier==0.5.6
PyYAML==6.0.1
PyYAML==6.0.2
regex==2023.12.25
Requests==2.32.3
rich==13.8.0
scikit_learn==1.4.1.post1
scipy==1.14.1
sentence_transformers==3.0.1
setuptools==68.2.2
torch==2.1.2
tqdm==4.66.4
transformers==4.40.2
typing_extensions==4.12.2
vertexai==1.49.0
voyageai==0.2.3
